requirements.txt
opencv-python>=4.8
numpy>=1.24
matplotlib>=3.7
torch>=2.1
torchvision>=0.16
pandas>=2.0
scikit-learn>=1.3
pillow>=10.0
mediapipe>=0.10

